refactor: shared.detail 패키지 내 컴포넌트 위치 이동 및 구조 정리
- `shared.detail.components` 패키지에 위치하던 공통 UI 컴포넌트들을 상위 패키지인 `shared.detail`로 이동
- 대상 컴포넌트: `DeleteConfirmDialog`, `EditDropdownMenu`, `InfoCard`, `ReceiversCard`
- 패키지 구조 변경에 따라 이를 참조하는 GalleryDetail, MemorialGuidelineDetail, SocialNetworkDetail 등 주요 화면의 임포트 경로 업데이트
- 관련 파일들의 패키지 선언부 수정 (`shared.detail.components` -> `shared.detail`)
