+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/detail/afternotedetail/GalleryDetailScreen.kt
@@ -40,11 +40,11 @@ import com.afternote.core.ui.theme.AfternoteDesign
 import com.afternote.core.ui.topbar.DetailTopBar
 import com.afternote.feature.afternote.presentation.R
 import com.afternote.feature.afternote.presentation.author.navigation.AfternoteLightTheme
+import com.afternote.feature.afternote.presentation.shared.detail.DeleteConfirmDialog
 import com.afternote.feature.afternote.presentation.shared.detail.DetailSection
+import com.afternote.feature.afternote.presentation.shared.detail.EditDropdownMenu
 import com.afternote.feature.afternote.presentation.shared.detail.MessageSection
 import com.afternote.feature.afternote.presentation.shared.detail.ProcessingMethodsSection
-import com.afternote.feature.afternote.presentation.shared.detail.components.DeleteConfirmDialog
-import com.afternote.feature.afternote.presentation.shared.detail.components.EditDropdownMenu
 import com.afternote.feature.afternote.presentation.shared.model.ReceiverUiModel
 
 /**
