+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/shared/detail/InfoCard.kt
@@ -1,4 +1,4 @@
-package com.afternote.feature.afternote.presentation.shared.detail.components
+package com.afternote.feature.afternote.presentation.shared.detail
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
